test: 예약 시간 인수테스트 분리
